[Fix] 리뷰사항 반영 2차
